feat: Split request_editor.rs into Feature Modules Without Behavior Changes.
